requirements.txt
gradio
pdfminer.six
python-docx
sentence-transformers
scikit-learn
transformers
torch
pillow
lxml
